Corrected minitests
