+++ tests/test_encapsulation.rb
@@ -14,6 +14,7 @@
 
 require_relative 'ciscotest'
 require_relative '../lib/cisco_node_utils/encapsulation'
+require_relative '../lib/cisco_node_utils/vdc'
 
 # TestEncapsulation - Minitest for Encapsulation node utility class
 class TestEncapsulation < CiscoTestCase
@@ -24,6 +25,8 @@ class TestEncapsulation < CiscoTestCase
     Encapsulation.encaps.each do |_encap, obj|
       obj.destroy
     end
+  rescue
+    skip('Unsupported in non F3 vdcs')
   end
 
   def setup
@@ -42,11 +45,13 @@ class TestEncapsulation < CiscoTestCase
   # TESTS
 
   def test_encapsulation_create_destroy
+    compatible_interface?
     encap = Encapsulation.new('cisco')
     encap.destroy
   end
 
   def test_encapsulation_dot1_mapping
+    compatible_interface?
     encap = Encapsulation.new('cisco')
     assert_equal(encap.default_dot1q_map, encap.dot1q_map,
                  'Error: dot1q is not matching')
@@ -61,6 +66,7 @@ class TestEncapsulation < CiscoTestCase
   end
 
   def test_invalid_range_dot1q_mapping
+    compatible_interface?
     encap = Encapsulation.new('cisco')
     assert_equal(encap.default_dot1q_map, encap.dot1q_map,
                  'Error: dot1q is not matching')
